+++ code/src/cljs/sixsq/nuvla/ui/main/components.cljs
@@ -68,7 +68,7 @@
 (defn SearchInput
   [opts]
   (let [tr (subscribe [::i18n-subs/tr])]
-    [ui/Input (merge {:placeholder (@tr [:search])
+    [ui/Input (merge {:placeholder (str (@tr [:search]) "...")
                       :icon        (when-not (:action opts) "search")}
                      opts)]))